requirements.txt
streamlit>=1.30.0
pytrends>=4.9.0
google-play-scraper>=1.2.7
google-api-python-client>=2.100.0
plotly>=5.18.0
pandas>=2.1.0
python-dotenv>=1.0.0
requests>=2.27.0
urllib3<2
beautifulsoup4>=4.12.0
diskcache>=5.6.0
